handle private channels

+++ bot.py
@@ -21,7 +21,7 @@ class Bot(object):
         self.emoji = ":book:"
         self.oauth = {"client_id": os.environ.get("CLIENT_ID"),
                       "client_secret": os.environ.get("CLIENT_SECRET"),
-                      "scope": "bot,channels:history"}
+                      "scope": "bot,channels:history,groups:history"}
         self.verification = os.environ.get("VERIFICATION_TOKEN")
         self.client = SlackClient("")
 
@@ -45,26 +45,30 @@ class Bot(object):
                                 client_secret=self.oauth["client_secret"],
                                 code=code
                                 )
-        print(auth_response)
-        team_id = auth_response["team_id"]
-        # To keep track of authorized teams and their associated OAuth tokens,
-        # we will save the team ID and bot tokens to Postgre
 
-        conn = psycopg2.connect(DATABASE_URL, sslmode='require')
-        cur = conn.cursor()
-        cur.execute("CREATE TABLE IF NOT EXISTS authed_teams (id serial PRIMARY KEY, team_id varchar, bot_token varchar(256));")
-        cur.execute(sql.SQL("select bot_token FROM authed_teams WHERE team_id = {}").format(sql.Literal(team_id)))
-        db_token = cur.fetchone()
+        if "team_id" in auth_response:
+            team_id = auth_response["team_id"]
+            # To keep track of authorized teams and their associated OAuth tokens,
+            # we will save the team ID and bot tokens to Postgre
 
-        if db_token is not None:
-            cur.execute(sql.SQL("DELETE FROM authed_teams WHERE team_id = {}").format(sql.Literal(team_id)))
+            conn = psycopg2.connect(DATABASE_URL, sslmode='require')
+            cur = conn.cursor()
+            cur.execute("CREATE TABLE IF NOT EXISTS authed_teams (id serial PRIMARY KEY, team_id varchar, bot_token varchar(256));")
+            cur.execute(sql.SQL("select bot_token FROM authed_teams WHERE team_id = {}").format(sql.Literal(team_id)))
+            db_token = cur.fetchone()
+
+            if db_token is not None:
+                cur.execute(sql.SQL("DELETE FROM authed_teams WHERE team_id = {}").format(sql.Literal(team_id)))
+                conn.commit()
+            db_token = auth_response["bot"]["bot_access_token"]
+            cur.execute(sql.SQL("INSERT INTO authed_teams (team_id, bot_token) VALUES ({}, {})").format(sql.Literal(team_id), sql.Literal(db_token)))
             conn.commit()
-        db_token = auth_response["bot"]["bot_access_token"]
-        cur.execute(sql.SQL("INSERT INTO authed_teams (team_id, bot_token) VALUES ({}, {})").format(sql.Literal(team_id), sql.Literal(db_token)))
-        conn.commit()
-        cur.close()
-        conn.close()
-        self.client = SlackClient(db_token)
+            cur.close()
+            conn.close()
+            self.client = SlackClient(db_token)
+        else:
+            print("Authentication failed!\n")
+            print(auth_response)
 
     def documentation_message(self, package, fun, channel):
         """
@@ -94,7 +98,6 @@ class Bot(object):
                                             icon_emoji=self.emoji,
                                             attachments=msg.attachments
                                             )
-        print(post_message)
 
     def update_client(self, team_id):
         """
